requirements.txt
django-organizations
# required by django-organizations:
django-extensions
# required by custom DRF Permission class:
djangorestframework
